Crate edit_distance

Edit distance

The Levenshtein edit distance between two strings is the number of individual single-character changes (insert, delete, substitute) necessary to change string a into b.

This can be a used to order search results, for fuzzy auto-completion, and to find candidates for spelling correction.

References

Wikipedia: Levenshtein distance
NIST: Levenshtein distance

Functions

edit_distance
Returns the edit distance between strings a and b.